+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.1.1
+# 0.1.1 [WIP]
 - WIP (ainda não disponível para uso em produção)
 - Implementada consulta de NFEs(V1) por id do gateway
 - Implementada consulta de NFEs(V1) por id externo
@@ -7,6 +7,8 @@
 - Implementado Habilitar Empresa
 - Implementado Desabilitar Empresa
 - Implementado Vincular Logotipo de Empresa
+- Implementado Vincular Certificado Digital de Empresa
+
 # 0.1.0
 - WIP (ainda não disponível para uso em produção)